backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-admin-inline-paginator==0.4.0
django-cors-headers==4.2.0
django-filter==23.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Levenshtein==0.23.0
MarkupSafe==2.1.3
numpy==1.23.5
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
psycopg2-binary==2.9.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
requests==2.31.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
